replace \$ with $ inside listings
Even though the listings package doesn't care about $ inside listings, Emacs
thinks everything between two is in LaTeX math mode and highlights it as
such. If I escape it, Emacs no longer thinks it's math mode, but listings prints
it as a literal "\$", but we can just replace it with a regular "$".